move selector location
