+++ app/selectors/tokenList.ts
@@ -1,5 +1,3 @@
-import { selectAssetsBySelectedAccountGroup as _selectAssetsBySelectedAccountGroup } from '@metamask/assets-controllers';
-import { Hex } from '@metamask/utils';
 import { createSelector } from 'reselect';
 import { selectTokenSortConfig } from './preferencesController';
 import { selectIsEvmNetworkSelected } from './multichainNetworkController';
@@ -73,47 +71,4 @@ const _selectSortedTokenKeys = createSelector(
 export const selectSortedTokenKeys = createDeepEqualSelector(
   _selectSortedTokenKeys,
   (keys) => keys.filter(({ address, chainId }) => address && chainId),
-);
-
-export const selectAssetsBySelectedAccountGroup = createDeepEqualSelector(
-  (state: RootState) => {
-    const {
-      AccountTreeController,
-      AccountsController,
-      TokensController,
-      TokenBalancesController,
-      TokenRatesController,
-      MultichainAssetsController,
-      MultichainBalancesController,
-      MultichainAssetsRatesController,
-      CurrencyRateController,
-      NetworkController,
-      AccountTrackerController,
-    } = state.engine.backgroundState;
-
-    return {
-      ...AccountTreeController,
-      ...AccountsController,
-      ...TokensController,
-      ...TokenBalancesController,
-      ...TokenRatesController,
-      ...MultichainAssetsController,
-      ...MultichainBalancesController,
-      ...MultichainAssetsRatesController,
-      ...CurrencyRateController,
-      ...NetworkController,
-      ...(AccountTrackerController as {
-        accountsByChainId: Record<
-          Hex,
-          Record<
-            Hex,
-            {
-              balance: Hex | null;
-            }
-          >
-        >;
-      }),
-    };
-  },
-  (filteredState) => _selectAssetsBySelectedAccountGroup(filteredState),
 );
